new file:   ModuleBase/user/action/myinfo.php

+++ ModuleBase/user/action/class_edit.php
@@ -11,27 +11,59 @@ if(isset($_REQUEST['id'])){
 		$id = intval($id);
 		$uc->setPrimaryKey($id);
 		if(isset($_REQUEST['delete'])){
-			$uc->destroy();
+			if($id > CUserDef::BANNED_DEL_MAX_CLASS_ID)
+				$uc->destroy();
 		}
 		else if(isset($_REQUEST['edit'])){
 			$list[] = $uc->get();
 		}
 		else if(isset($_REQUEST['edit_submit'])){
+			if(empty($_REQUEST['name'][$k]) || empty($_REQUEST['code'][$k])){
+				$error[] = $k.'#'.$mbs_appenv->lang('miss_args');
+				continue;
+			}
 			try {
 				$ret = $uc->set(array(
 					'name' => $_REQUEST['name'][$k],
 					'code' => $_REQUEST['code'][$k]
 				));
+				if(!$ret){
+					$error[] = $mbs_appenv->lang(array('name', 'or', 'class_code', 'existed'))
+						.':'.$_REQUEST['name'][$k].':'.$_REQUEST['code'][$k];
+				}
 				$list[] = array('id'=>$id, 'name'=>$_REQUEST['name'][$k], 'code'=>$_REQUEST['code'][$k]);
 			} catch (Exception $e) {
 				$error[] = $e->getMessage();
 			}
+			
 		}
 	}
 	
 	if(isset($_REQUEST['delete'])){
 		$mbs_appenv->echoex($mbs_appenv->lang('operation_success', 'common'), '', $mbs_appenv->toURL('class'));
 		exit(0);
+	}
+}else{
+	$list = array(array('name'=>'', 'code'=>''));
+	if(isset($_REQUEST['edit_submit'])){
+		if(empty($_REQUEST['name'][0]) || empty($_REQUEST['code'][0])){
+			$error[] = $mbs_appenv->lang('miss_args');
+		}
+		if(empty($error)){
+			try {
+				$ret = $uc->add(array(
+					'name' => $_REQUEST['name'][0],
+					'code' => $_REQUEST['code'][0],
+					'create_time' => time()
+				));
+			} catch (Exception $e) {
+				$error[] = $e->getMessage();
+			}
+			if(!$ret){
+				$error[] = $mbs_appenv->lang(array('name', 'or', 'class_code', 'existed'))
+					.':'.$_REQUEST['name'][0].':'.$_REQUEST['code'][0];;
+			}
+		}
 	}
 }
 
@@ -40,50 +72,65 @@ if(isset($_REQUEST['id'])){
 <html>
 <head>
 <title><?php mbs_title()?></title>
-<link href="<?php echo $mbs_appenv->sURL('pure-min.css')?>" rel="stylesheet">
-<link href="<?php echo $mbs_appenv->sURL('core.css')?>" rel="stylesheet">
+<meta charset="UTF-8">
+<meta name="viewport" content="width=device-width,initial-scale=1.0,user-scalable=no,minimum-scale=1.0,maximum-scale=1.0">
+<title><?php mbs_title()?></title>
+<!--[if lt ie 9]>
+	<script>
+		document.createElement("article");
+		document.createElement("section");
+		document.createElement("aside");
+		document.createElement("footer");
+		document.createElement("header");
+		document.createElement("nav");
+</script>
+<![endif]-->
+<link rel="stylesheet" href="<?php echo $mbs_appenv->sURL('reset.css')?>" type="text/css" />
+<link rel="stylesheet" href="<?php echo $mbs_appenv->sURL('global.css')?>" />
+<link rel="stylesheet" href="<?php echo $mbs_appenv->sURL('userManager.css')?>">
+<style type="text/css">
+.even{background-color:#eee;}
+</style>
 </head>
 <body>
-<div class=header><?php echo $mbs_appenv->lang('header_html', 'common')?></div>
-<div class="pure-g wrapper">
-    <div class="pure-u-1">
-    	<?php if(!empty($error)){ ?>
-		<div class=error><?php  foreach($error as $e){?><p><?php echo CStrTools::txt2html($e)?></p><?php }?>
-		<a href="#" class=close onclick="this.parentNode.parentNode.removeChild(this.parentNode)" >&times;</a>
-		</div>
-		<?php }else if(isset($_REQUEST['edit_submit'])){ ?>
-		<div class=success><?php echo $mbs_appenv->lang('operation_success', 'common')?></div>
-		<?php }?>
-		
-		<form class="pure-form" method="post">
-        	<h3><?php echo $mbs_appenv->lang('edit')?>
-        		<a class=back href="<?php echo $mbs_appenv->toURL('class', '')?>">&lt;<?php echo $mbs_appenv->lang('back', 'common')?></a>
-        	</h3>
-			<table class="pure-table" style="width: 100%;margin:1em 0;">
-			    <thead>
-			        <tr>
-			            <th><?php echo $mbs_appenv->lang('class_name')?></th>
-			            <th><?php echo $mbs_appenv->lang('class_code')?></th>
-			        </tr>
-			    </thead>
-			
-			    <tbody>
-			    <?php $k = -1; foreach($list as $k=>$row){?>
-			    	<input type="hidden" name="id[]" value="<?php echo $row['id']?>" />
-			        <tr <?php echo 1 == $k%2 ? 'class=pure-table-odd':'' ?>>
-			            <td><input type="text" name="name[]" value="<?php echo $row['name']?>" /></td>
-			            <td><input type="text" name="code[]" value="<?php echo $row['code']?>" /></td>
-			        </tr>
-			    <?php }?>
-			    </tbody>
-			</table>
-			<?php if(-1 == $k){ echo '<p class=no-data>', $mbs_appenv->lang('no_data', 'common'), '</p>'; 
-    		}else{ ?>
-    		<button class="pure-button pure-button-primary" name="edit_submit" type="submit"><?php echo $mbs_appenv->lang('edit')?></button>
-    		<?php }?>
-		</form>
+<div class="userManager">
+	<h2 class="tit"><?php echo $mbs_appenv->lang(array('class', 'edit'))?>
+    	<a href="<?php echo $mbs_appenv->toURL('class')?>" class="btn-create"><span class="back-icon"></span><?php echo $mbs_appenv->lang('back')?></a>
+    </h2>
+    
+    <?php if(!empty($error)){ ?>
+	<div class=error><?php  foreach($error as $e){?><p><?php echo CStrTools::txt2html($e)?></p><?php }?>
+	<a href="#" class=close onclick="this.parentNode.parentNode.removeChild(this.parentNode)" >&times;</a>
+	</div>
+	<?php }else if(isset($_REQUEST['edit_submit'])){ ?>
+	<div class=success><?php echo $mbs_appenv->lang('operation_success', 'common')?></div>
+	<?php }?>
+	
+	<form method="post" name=_form action="<?php echo $mbs_appenv->toURL('class_edit')?>">
+	<?php foreach($list as $k=>$row){ if(isset($row['id'])){ ?>
+	<input type="hidden" name="id[]" value="<?php echo $row['id']?>" />
+	<?php }?>
+	<div <?php echo 1==$k%2 ? 'class="even"' : ''?>>
+	<div class="inpBox mb17">
+   		<label for="name" class="labelL"><?php echo $mbs_appenv->lang('name')?>&nbsp;:&nbsp;</label>
+	    <input id="name" class="inpTit" name="name[]" type="text" value="<?php echo $row['name']?>" 
+	    	placeholder="<?php echo $mbs_appenv->lang('please_input')?>" required />
     </div>
+    <div class="inpBox mb17">
+    	<label for="code" class="labelL"><?php echo $mbs_appenv->lang('class_code')?>&nbsp;:&nbsp;</label>
+    	<input id="code" class="inpTit" name="code[]" type="text" value="<?php echo $row['code']?>" 
+		    	placeholder="<?php echo $mbs_appenv->lang('please_input')?>" required />
+    </div>
+    </div>
+    <?php } ?>
+    <div class="btnBox">
+    	<input type="hidden" name="edit_submit" value="" />
+    	<label for="code" class="labelL">&nbsp;</label>
+        <a href="javascript:document._form.submit();" class="btn-send"><?php echo $mbs_appenv->lang('submit')?></a>
+        <a href="<?php echo $mbs_appenv->toURL('class')?>" class="btn-cancle"><?php echo $mbs_appenv->lang('cancel')?></a>
+    </div>
+    </form>
+    
 </div>
-<div class=footer></div>
 </body>
 </html>
